+++ src/utils/index.js
@@ -1,5 +1,6 @@
 const flattenObject = require('./flattenObject')
 const interpolateContext = require('./interpolateContext')
 const fromPaths = require('./fromPaths')
+const template = require('./template')
 
-module.exports = { flattenObject, fromPaths, interpolateContext }
+module.exports = { flattenObject, fromPaths, interpolateContext, template }