+++ src/utils/interpolateContext.js
@@ -1,7 +1,7 @@
 const R = require('ramda')
-const { pope } = require('pope')
+const template = require('./template')
 
-const interpolateValue = context => value => pope(value, context)
+const interpolateValue = context => value => template(value, context)
 
 module.exports = (rules, context) => {
   return R.map(interpolateValue(context), rules)