+++ test/utils/interpolateContext.spec.js
@@ -1,5 +1,6 @@
 const chai = require('chai')
 const utils = require('../../src/utils')
+const Macroable = require('macroable')
 
 const { expect } = chai
 
@@ -32,6 +33,26 @@ describe('utils | interpolateContext', () => {
 
     expect(utils.interpolateContext(rules, context)).to.deep.equal({
       name: 'unique:users,email,id,1'
+    })
+  })
+
+  it('interpolate context from macroable', () => {
+    class Query extends Macroable {
+    }
+
+    Query._getters = {}
+    Query.getter('field', () => 'id')
+
+    const rules = {
+      name: 'unique:users,email,{{query.field}}'
+    }
+
+    const context = {
+      query: new Query()
+    }
+
+    expect(utils.interpolateContext(rules, context)).to.deep.equal({
+      name: 'unique:users,email,id'
     })
   })
 })