+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 
+### Fixed
+- Allow to interpolate values from Macroable
+
+### Removed
+- Dependency to pope library
+
 ## [1.0.0] - 2017-10-22
 ### Added
 - Rule flattener